+++ .idea/Spring_boot_external_courses.iml
@@ -6,4 +6,7 @@
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
+  <component name="SonarLintModuleSettings">
+    <option name="uniqueId" value="584c9496-94f9-490f-bef6-c148b8b18bb2" />
+  </component>
 </module>
