+++ Servlets/ServletInit/src/main/java/Servlet.java
@@ -1,0 +1,23 @@
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.*;
+import java.io.IOException;
+import java.io.PrintWriter;
+
+@WebServlet("/servlet")
+public class Servlet extends HttpServlet{
+
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        response.setContentType("text/html;charset=UTF-8");
+        PrintWriter out = response.getWriter();
+
+        out.print("<head>");
+        out.print("<title>Resultado Servlet</title>");
+        out.print("</head>");
+        out.print("<body>");
+
+        out.print("Hi");
+
+        out.print("</body>");
+        out.print("</html>");
+    }
+}
