--- Servlets/ServletsLearning/src/Main.java
@@ -1,5 +1,0 @@
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
