--- Servlets/ServletsLearning/src/servlets/servletMain.java
@@ -1,4 +1,0 @@
-package servlets;
-
-public class servletMain {
-}
